* Tile changes from OT, yet Static/Dynamic, etc.
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@1712 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ fileloader.cpp
@@ -379,7 +379,7 @@ inline bool FileLoader::readBytes(uint8_t* buffer, int32_t size, int32_t pos)
 			m_cache_offset = pos - m_cached_data[i].base;
 
 			//get maximum read block size and calculate remaining bytes
-			reading = std::min(remain, m_cached_data[i].size - m_cache_offset);
+			reading = std::min(remain, (uint32_t)m_cached_data[i].size - m_cache_offset);
 			remain = remain - reading;
 
 			//read it
