+++ luascript.cpp
@@ -2422,7 +2422,7 @@ int32_t LuaScriptInterface::internalGetPlayerInfo(lua_State* L, PlayerInfo_t inf
 				value = player->getSoul();
 				break;
 			case PlayerInfoFreeCap:
-				value = player->getFreeCapacity();
+				value = (int64_t)player->getFreeCapacity();
 				break;
 			case PlayerInfoGuildId:
 				value = player->getGuildId();
